pyproject.toml
[project]
name = "intigriti-mcp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",
]

[project.scripts]
intigriti-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["server.py", "auth.py", "client.py", "cache.py"]
